restructure and rename source files
