+++ Sources/StringsCatalogTranslate/Keys/Keychain.swift
@@ -1,0 +1,7 @@
+import Foundation
+import KeychainSimple
+
+enum Keychain {
+    static let namePrefix = "tools.xcode.translate_strings"
+    static let access = KeychainAccess(itemNamePrefix: namePrefix)
+}
